Clean up and simplify ant/gradle builds and README
Automatically uses protobuf jar and SVMPProtocol.java from the
svmp-protocol-def project that is assumed to be checked out in a
directory parallel to the test-proxy.

Gradle now uses svmp-protocol-def as as subproject and doesn't need
any file copying tricks or ant custom rules to generate code straight
from the .proto definition file.

Change-Id: I605d932587ff0aee09f5961853f3d55d89827813